# api.yaml
openapi: 3.0.3
info:
  title: Placement Park API
  description: REST API for the Placement Park placement management portal.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
...
paths:
...
  /analytics/admin/monthly-trends:
    get:
      tags: [Analytics]
      summary: Get monthly application trends
      parameters:
        - $ref: '#/components/parameters/Year'
      responses:
        '200': {description: Monthly trend data}
...
  /analytics/admin/branch-analytics:
    get:
      tags: [Analytics]
      summary: Get branch statistics
      parameters:
        - $ref: '#/components/parameters/Year'
      responses:
        '200': {description: Branch statistics}
  /analytics/admin/package-analytics:
    get:
      tags: [Analytics]
      summary: Get package statistics
      parameters:
        - $ref: '#/components/parameters/Year'
      responses:
        '200': {description: Package statistics}
  /analytics/admin/drive-performance:
    get:
      tags: [Analytics]
      summary: Get drive performance
      parameters:
        - $ref: '#/components/parameters/Year'
      responses:
        '200': {description: Drive performance data}
  /analytics/admin/recent-activities:
    get:
      tags: [Analytics]
      summary: Get recent activities
      parameters:
        - name: limit
          in: query
          schema: {type: integer, default: 20}
      responses:
        '200': {description: Recent activities}
  /analytics/admin/insights:
    get:
      tags: [Analytics]
      summary: Get admin insights
      parameters:
        - $ref: '#/components/parameters/Year'
      responses:
        '200': {description: Admin insights}
...
components:
  parameters:
...
    Year:
      name: year
      in: query
      schema: {type: integer}
...